feat: Update Product Manager so that it can update and remove products.
Added path http://localhost:3000/:id/edit which includes an auto-filled form to edit a given product.

Added a delete button to the List component that will remove the product from the database.

After removing the product from the Detail component, it will redirect to the main view

After removing the product from the main view, it updates the DOM so that the product is removed.